Field deletions and Material Demand Calculation

+++ models/materials_per_product.py
@@ -4,7 +4,7 @@ class MaterialsPerProduct(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     material_id = db.Column(db.Integer, db.ForeignKey('material.id'), nullable=False)
     product_id = db.Column(db.Integer, db.ForeignKey('product.id'), nullable=False)
-    amount = db.Column(db.Integer, nullable=False)
+    amount = db.Column(db.Float, nullable=False)
     
     def serialize(self):
         return {
